Move _LINK
